make "internal-configuration" as part of brokers
