--- pulsar-client-tools/src/main/java/org/apache/pulsar/admin/cli/CmdInternalConfiguration.java
@@ -1,41 +1,0 @@
-/**
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package org.apache.pulsar.admin.cli;
-
-import com.beust.jcommander.Parameters;
-import org.apache.pulsar.client.admin.PulsarAdmin;
-
-@Parameters(commandDescription = "Operations about internal configuration")
-public class CmdInternalConfiguration extends CmdBase {
-
-    @Parameters(commandDescription = "Get internal configuration information")
-    private class Get extends CliCommand {
-
-        @Override
-        void run() throws Exception {
-            print(admin.internalConfiguration().getInternalConfigurationData());
-        }
-
-    }
-
-    public CmdInternalConfiguration(PulsarAdmin admin) {
-        super("internal-configuration", admin);
-        jcommander.addCommand("get", new Get());
-    }
-}
